Olinda v1: Distiller for ZairaChem

+++ src/olinda/distillation.py
@@ -1,8 +1,12 @@
 """Distillation module."""
 
+from warnings import filterwarnings
+filterwarnings(action="ignore")
+
 import os
 from pathlib import Path
 import shutil
+import math
 from typing import Any, Optional
 
 from cbor2 import dump
@@ -11,33 +15,37 @@ import pytorch_lightning as pl
 import torch
 from tqdm import tqdm
 
+import boto3
+from botocore import UNSIGNED
+from botocore.config import Config
+import zipfile
+
 import tensorflow as tf
 import tf2onnx
 import onnx
+import pandas as pd
 
 from olinda.data import ReferenceSmilesDM, FeaturizedSmilesDM, GenericOutputDM
 from olinda.featurizer import Featurizer, MorganFeaturizer, Flat2Grid
 from olinda.generic_model import GenericModel
 from olinda.tuner import ModelTuner, KerasTuner
-from olinda.utils import calculate_cbor_size, get_workspace_path
+from olinda.utils.utils import calculate_cbor_size, get_workspace_path
+from olinda.utils.s3 import ProgressPercentage
 
-
-def distill(
-    model: Any,
-    working_dir: Path = get_workspace_path(),
-    featurizer: Optional[Featurizer] = MorganFeaturizer(),
-    clean: bool = False,
-    tuner: ModelTuner = KerasTuner([1, 3]),
-    reference_smiles_dm: Optional[ReferenceSmilesDM] = None,
-    featurized_smiles_dm: Optional[FeaturizedSmilesDM] = None,
-    generic_output_dm: Optional[GenericOutputDM] = None,
-    test: bool = False,
-    num_data: int = 1999380,
-) -> pl.LightningModule:
-    """Distill models.
-
-    Args:
-        model (Any): Teacher Model.
+### TODO: Improve object-oriented setup of distillation code segments
+class Distiller(object):
+    def __init__(self,
+        featurizer: Optional[Featurizer] = MorganFeaturizer(),
+        tuner: ModelTuner = KerasTuner(),
+        reference_smiles_dm: Optional[ReferenceSmilesDM] = None,
+        featurized_smiles_dm: Optional[FeaturizedSmilesDM] = None,
+        generic_output_dm: Optional[GenericOutputDM] = None,
+        num_data: int = 100000,
+        clean: bool = True,
+        test: bool = False,
+    ):
+        """
+        Args:
         featurizer (Optional[Featurizer]): Featurizer to use.
         working_dir (Path): Path to model workspace directory.
         clean (bool): Clean workspace before starting.
@@ -46,41 +54,125 @@ def distill(
         featurized_smiles_dm (Optional[FeaturizedSmilesDM]): Reference Featurized SMILES datamodules.
         generic_output_dm (Optional[GenericOutputDM]): Precalculated training dataset for student model.
         test (bool): Run a test distillation on a smaller fraction of the dataset.
-        num_data: (int) : Set the number of ChEMBL training points to use (up to 1999380)
-    Returns:
-        pl.LightningModule: Student Model.
-    """
+        """     
+        self.working_dir = get_workspace_path()
+        self.featurizer = featurizer
+        self.tuner = tuner
+        self.reference_smiles_dm = reference_smiles_dm
+        self.featurized_smiles_dm = featurized_smiles_dm
+        self.generic_output_dm = generic_output_dm
+        self.num_data = num_data
+        if test:
+            self.num_data = self.num_data // 10
+        self.clean = clean
+        self.test = test      
 
-    # Convert model to a generic model
-    model = GenericModel(model)
-    student_training_dm = generic_output_dm
-    if student_training_dm is None:
-        # Prepare reference smiles datamodule
-        if reference_smiles_dm is None:
-            reference_smiles_dm = ReferenceSmilesDM(num_data=num_data)
-        reference_smiles_dm.prepare_data()
-        if not test:
-            reference_smiles_dm.setup("train")
+    def distill(self, model: Any) -> pl.LightningModule:
+        """Distill models.
+        
+        Args:
+            model (Any): Teacher Model.
+        Returns:
+            pl.LightningModule: Student Model.
+        """
+        
+        if self.clean is True:
+            clean_workspace(Path(self.working_dir), reference=True)
+        
+        # Convert model to a generic model
+        model = GenericModel(model)
+        if model.type == "zairachem":
+            fetch_ref_library()
+            ref_library = os.path.join(os.path.expanduser("~"), "olinda", "precalculated_descriptors", "olinda_reference_library.csv")
+            precalc_smiles_df = pd.read_csv(ref_library, header=None)
+            ref_data = len(precalc_smiles_df)
+            self.reference_smiles_dm = ReferenceSmilesDM(num_data=self.num_data)
+            self.reference_smiles_dm.prepare_data()
+            self.reference_smiles_dm.setup("train")
+            
+            if self.num_data > ref_data:
+                self.num_data = ref_data  
+            zairachem_folds = math.ceil(self.num_data / 50000)
+            fetch_descriptors(zairachem_folds)
+            
+            self.featurized_smiles_dm = gen_featurized_smiles(self.reference_smiles_dm, self.featurizer, self.working_dir, num_data=self.num_data, clean=self.clean)
+            self.featurized_smiles_dm.setup("train")       
+            student_training_dm = gen_model_output(model, self.featurized_smiles_dm, self.working_dir, self.num_data, self.clean)
         else:
-            reference_smiles_dm.setup("val")
+            student_training_dm = self.generic_output_dm
+            
+        if student_training_dm is None:
+            # Prepare reference smiles datamodule
+            if reference_smiles_dm is None:
+                self.reference_smiles_dm = ReferenceSmilesDM(num_data=self.num_data)
+            self.reference_smiles_dm.prepare_data()
+            self.reference_smiles_dm.setup("train")
+        
+            # Generate student model training dataset
+            student_training_dm = gen_training_dataset(
+                model,
+                self.featurizer,
+                self.reference_smiles_dm,
+                self.featurized_smiles_dm,
+                self.working_dir,
+                self.num_data,
+                self.clean,
+            )
+        
+        # Select and Train student model
+        student_model = self.tuner.fit(student_training_dm)
+        model_onnx = convert_to_onnx(student_model, self.featurizer)
+    
+        return model_onnx
 
-        # Generate student model training dataset
-        student_training_dm = gen_training_dataset(
-            model,
-            featurizer,
-            reference_smiles_dm,
-            featurized_smiles_dm,
-            working_dir,
-            num_data,
-            clean,
-        )
+def fetch_ref_library():
+    s3 = boto3.resource('s3', config=Config(signature_version=UNSIGNED))
+    bucket = s3.Bucket('olinda')
+    
+    path = os.path.join(os.path.expanduser("~"), "olinda", "precalculated_descriptors")
+    os.makedirs(path, exist_ok=True)
+    
+    lib_name = "olinda_reference_library.csv"
+    ref_lib = os.path.join(path, lib_name)
+    # if no reference library or the size differs to the S3 bucket version
+    if os.path.exists(ref_lib) == False or bucket.Object(key=lib_name).content_length != os.path.getsize(ref_lib):
+        if os.path.exists(ref_lib):
+            os.remove(ref_lib)
+        bucket.download_file(
+                "olinda_reference_library.csv", ref_lib,
+                Callback=ProgressPercentage(bucket, "olinda_reference_library.csv")
+                )
 
-    # Select and Train student model
-    student_model = tuner.fit(student_training_dm)
-    model_onnx = convert_to_onnx(student_model, featurizer)
-
-    return model_onnx
-
+def fetch_descriptors(
+    num_folds: int
+    ):
+    """Check if required precalculated descriptor folds are on disk and fetch missing folds 
+    
+    Args:
+        num_folds (int): Number of folds of 50k precalculated descriptors
+    """
+    
+    s3 = boto3.resource('s3', config=Config(signature_version=UNSIGNED))
+    bucket = s3.Bucket('olinda')
+    
+    path = os.path.join(os.path.expanduser("~"), "olinda", "precalculated_descriptors")
+    
+    for i in range(num_folds):
+        fold = "olinda_reference_descriptors_" + str(i*50) + "_" + str((i+1)*50) + "k"
+        dest = os.path.join(path, fold)
+        fold_zip = fold + ".zip"
+        dest_zip = dest + ".zip"
+        
+        if os.path.exists(dest) == False:
+            print("Downloading precalculated descriptors: fold " + str(i+1))
+            bucket.download_file(
+                fold_zip, dest_zip,
+                Callback=ProgressPercentage(bucket, fold_zip)
+                )
+            with zipfile.ZipFile(dest_zip, 'r') as zip_ref:
+                zip_ref.extractall(path)
+            assert os.path.exists(dest)
+            os.remove(dest_zip)
 
 def gen_training_dataset(
     model: pl.LightningModule,
@@ -183,22 +275,25 @@ def gen_featurized_smiles(
                 continue   
             if i >= reference_smiles_dl.length:
                 break
+            if i >= num_data:
+                break
                      
             output = featurizer.featurize(batch[1])
             for j, elem in enumerate(batch[0]):
                 dump((elem.tolist(), batch[1][j], output[j].tolist()), feature_stream)
 
-    featurized_smiles_dm = FeaturizedSmilesDM(Path(working_dir), featurizer)
+    featurized_smiles_dm = FeaturizedSmilesDM(Path(working_dir), featurizer, num_data=num_data)
     
     return featurized_smiles_dm
 
-
 def gen_model_output(
-    featurized_smiles_dm: pl.LightningDataModule,
     model: GenericModel,
+    featurized_smiles_dm: pl.LightningDataModule,
     working_dir: Path,
+    ref_size: int,
     clean: bool = False,
-) -> onnx.onnx_ml_pb2.ModelProto:
+) -> pl.LightningDataModule:
+
     """Generate featurized smiles representation dataset.
 
     Args:
@@ -210,7 +305,8 @@ def gen_model_output(
     Returns:
         pl.LightningDataModule: Dateset with featurized smiles.
     """
-    os.makedirs(Path(working_dir) / model.name, exist_ok=True)
+
+    os.makedirs(os.path.join(working_dir, model.name), exist_ok=True)
     if clean is True:
         clean_workspace(Path(working_dir), model=model)
         featurized_smiles_dl = featurized_smiles_dm.train_dataloader()
@@ -241,6 +337,55 @@ def gen_model_output(
     with open(
         Path(working_dir) / (model.name) / "model_output.cbor", "wb"
     ) as output_stream:
+    
+        if model.type == "zairachem":
+            from loguru import logger
+            output = pd.DataFrame(columns = ["smiles", 'pred'])
+            for i in range(math.ceil(ref_size/50000)):
+                logger.info("Getting ZairaChem predictions for fold " + str(i+1) + " of " + str(math.ceil(ref_size/50000)))
+                folder = os.path.join(os.path.expanduser("~"), "olinda", "precalculated_descriptors", "olinda_reference_descriptors_" + str(i*50) + "_" + str((i+1)*50) + "k")
+                smiles_input_path = os.path.join(os.path.expanduser("~"), "olinda", "precalculated_descriptors", folder, "reference_library.csv")
+                preds = model(smiles_input_path)
+                output = pd.concat([output, preds])    
+            output = output[["smiles", "pred"]]   
+             
+            # save to zairachem model folder
+            zaira_distill_path = os.path.join(model.name[len(model.type)+1:], "distill") #get model root
+            if os.path.exists(zaira_distill_path) == False:
+                os.mkdir(zaira_distill_path)
+            output.to_csv(os.path.join(zaira_distill_path, "reference_predictions.csv"), index=False)
+            
+            # correct wrong zairachem predictions before training Olinda
+            training_output = model.get_training_preds()
+            for i, row in enumerate(training_output.iterrows()):
+                if row[1]["pred"] >= 0.0 and row[1]["pred"] <= 0.5 and row[1]["true"] == 1.0:
+                    training_output.at[i, "pred"] = 1.0 - row[1]["pred"]
+                elif row[1]["pred"] >= 0.5 and row[1]["pred"] <= 1.0 and row[1]["true"] == 0:
+                    training_output.at[i, "pred"] = 1.0 - row[1]["pred"]
+                    
+            """
+            # weight by data source: training/reference
+            # inverse of proportion of training compounds to all compounds
+            train_weight = 1 #round((training_output.shape[0] + ref_size) / len(training_output), 2)
+            """
+            
+            # inverse of ratio of predicted active to inactive 
+            y_bin_train = [1 if val > 0.5 else 0 for val in training_output["pred"]]
+            y_bin_ref = [1 if val > 0.5 else 0 for val in output["pred"]]
+            active_weight = (y_bin_train.count(0) + y_bin_ref.count(0)) / (y_bin_train.count(1) + y_bin_ref.count(1))
+            
+            print("Creating model prediction files")
+            train_counter = 0
+            morganFeat = MorganFeaturizer()
+            for i, row in training_output.dropna().iterrows():
+                fp = morganFeat.featurize([row["smiles"]])
+                if fp is None:
+                    continue
+                weight = active_weight
+                dump((i, row["smiles"], fp[0].tolist(), [row["pred"]], [weight]), output_stream)
+                train_counter += 1        
+            
+        ref_counter = 0        
         for i, batch in tqdm(
             enumerate(iter(featurized_smiles_dl)),
             total=featurized_smiles_dl.length,
@@ -248,8 +393,25 @@ def gen_model_output(
         ):
             if i < stop_step // len(batch[0]):
                 continue
-               
-            if model.type == "ersilia":
+
+            if model.type == "zairachem":
+                # final dataset a multiple of batch
+                combined_count = train_counter + featurized_smiles_dl.length*len(batch[0])
+                target_count = combined_count // len(batch[0]) * len(batch[0]) 
+                
+                for j, elem in enumerate(batch[1]):
+                    if ref_counter + train_counter == target_count:
+                        break
+                    if not output[output["smiles"] == elem].empty:
+                        pred_val = output[output["smiles"] == elem]["pred"].iloc[0]
+                        if pred_val > 0.5:
+                            weight = active_weight
+                        else:
+                            weight = 1
+                        dump((j, elem, batch[2][j], [pred_val], [weight]), output_stream)
+                        ref_counter += 1
+                
+            elif model.type == "ersilia":
                 output = model(batch[1])
                 for j, elem in enumerate(batch[1]):
                     dump((j, elem, batch[2][j], [output[j].tolist()]), output_stream)
@@ -259,9 +421,16 @@ def gen_model_output(
             	for j, elem in enumerate(batch[1]):
             	    dump((j, elem, batch[2][j], output[j].tolist()), output_stream)
 
-    model_output_dm = GenericOutputDM(Path(working_dir / (model.name)))
-    return model_output_dm
+        # Remove zairachem folder
+        if os.path.exists(os.path.join(get_workspace_path(), "zairachem_output_dir")):
+            shutil.rmtree(os.path.join(get_workspace_path(), "zairachem_output_dir"))
 
+    if model.type == "zairachem":
+        model_output_dm = GenericOutputDM(Path(working_dir / (model.name)), zaira_training_size = training_output.shape[0])
+    else:
+        model_output_dm = GenericOutputDM(Path(working_dir / (model.name)))   
+        
+    return model_output_dm
 
 def convert_to_onnx(
     model: pl.LightningModule,
@@ -286,7 +455,7 @@ def convert_to_onnx(
 
 
 def clean_workspace(
-    working_dir: Path, model: GenericModel = None, featurizer: Featurizer = None
+    working_dir: Path, model: GenericModel = None, featurizer: Featurizer = None, reference: bool = False
 ) -> None:
     """Clean workspace.
 
@@ -295,15 +464,25 @@ def clean_workspace(
         model (GenericModel): Wrapped Teacher model.
         featurizer (Featurizer): Featurizer to use.
     """
-
+    
+    curr_ref_smiles_path = Path(working_dir) / "reference" / "reference_smiles.csv"
+    orig_ref_smiles_path = os.path.join(os.path.expanduser("~"), "olinda", "precalculated_descriptors", "olinda_reference_library.csv")
+    
     if model:
         shutil.rmtree(Path(working_dir) / (model.name), ignore_errors=True)
         os.makedirs(Path(working_dir) / (model.name), exist_ok=True)
 
-    if featurizer:
+    if featurizer and os.path.exists(Path(working_dir) / "reference" / "reference_smiles_dl.joblib"):
         os.remove(Path(working_dir) / "reference" / "reference_smiles_dl.joblib")
-        os.remove(
-            Path(working_dir)
-            / "reference"
-            / f"featurized_smiles_{type(featurizer).__name__.lower()}.cbor"
+        os.remove(Path(working_dir) / "reference" / f"featurized_smiles_{type(featurizer).__name__.lower()}.cbor"
         )
+    
+    if reference and os.path.exists(curr_ref_smiles_path):
+        if os.path.exists(orig_ref_smiles_path):
+            curr_df = pd.read_csv(curr_ref_smiles_path, header=None, names=["SMILES"])
+            orig_df = pd.read_csv(orig_ref_smiles_path)
+            if not curr_df.equals(orig_df):
+                shutil.rmtree(Path(working_dir) / "reference")
+        else:
+            shutil.rmtree(Path(working_dir) / "reference")
+
